Shift countdown logic to parent component
